Added two pane mode for tablet
