+++ app/src/main/java/com/chichkanov/yandex_weather/di/AppComponent.java
@@ -14,7 +14,7 @@ import com.chichkanov.yandex_weather.interactor.WeatherInteractorImpl;
 import com.chichkanov.yandex_weather.repository.CityRepositoryImpl;
 import com.chichkanov.yandex_weather.repository.WeatherRepositoryImpl;
 import com.chichkanov.yandex_weather.ui.change_city.ChangeCityPresenter;
-import com.chichkanov.yandex_weather.ui.main.MainPresenter;
+import com.chichkanov.yandex_weather.ui.favorite_cities.FavoriteCitiesPresenter;
 import com.chichkanov.yandex_weather.ui.weather.WeatherPresenter;
 import com.chichkanov.yandex_weather.utils.Settings;
 
@@ -49,7 +49,7 @@ public interface AppComponent {
 
     WeatherPresenter getWeatherPresenter();
 
-    MainPresenter getMainPresenter();
+    FavoriteCitiesPresenter getFavoritesPresenter();
 
     void inject(WeatherDatabase weatherDatabase);
 }
