+++ app/src/main/java/com/chichkanov/yandex_weather/ui/main/OnDrawerEnabled.java
@@ -3,4 +3,6 @@ package com.chichkanov.yandex_weather.ui.main;
 
 public interface OnDrawerEnabled {
     void setDrawerEnabled(boolean enabled);
+
+    void closeDrawer();
 }
