+++ app/src/main/res/layout/fragment_favorite_cities.xml
@@ -1,0 +1,16 @@
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context="com.chichkanov.yandex_weather.ui.favorite_cities.FavoriteCitiesFragment">
+
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/rv_cities"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="@color/colorDrawer"
+        android:fitsSystemWindows="true"
+        app:itemBackground="@android:color/transparent" />
+
+</FrameLayout>
